Add files via upload
What is supervised learning?
Given a set of data with target column included, we want to train a model that can learn to map the input features (also known as the independent variables) to the target.

Machine Learning Models
We are going to focus on 2 types of ML models in this session. Both of them are supervised in nature (labels are known and are to be predicted).

Classification A classification model predicts discrete values. For example, classification models make predictions that answer questions like the following:

Is a given email message spam or not spam?

Is this an image of a dog, a cat, or a hamster?

Regression A regression model predicts continuous values. For example, regression models make predictions that answer questions like the following:

What is the value of a house in California?

What is the probability that a user will click on this ad?